requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
psycopg2-binary
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.4.0
gunicorn==21.2.0